+++ astroquery/casda/tests/__init__.py
@@ -1,0 +1,4 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""
+CSIRO ASKAP Science Data Archive (CASDA)
+"""